remove layer='foreground' attribute; unnecessary and counter-productive in this case
when a DS source defines a 'layer' attribute, ufo2ft assumes that that particular source is a sparse layer that only contributes glyph outlines, not any other piece of metadata (no fontinfo, no kerning, no groups) which are tied to the whole UFO and its default layer (default layer is the one pointing at the directory named glyphs, it doesn't matter it's called 'foreground' or 'public.default' or 'foobar').

In this case, 'foreground' happens to be the default layer for this source UFO so the layer attribute is unneeded; furthermore, it prevents the sparse kerning.plist data from being considered, so it's best to remove it.